+++ tests/integration_test.rs
@@ -8,7 +8,7 @@ fn create_heic_context() {
     unsafe {
         lh::heif_init(ptr::null_mut());
         let version = CStr::from_ptr(lh::heif_get_version()).to_string_lossy();
-        assert_eq!(version, "1.18.2");
+        assert_eq!(version, "1.19.5");
 
         let ctx = lh::heif_context_alloc();
         assert!(!ctx.is_null());